+++ crmeb-common/src/main/java/com/cshy/common/model/response/StoreOrderInfoResponse.java
@@ -98,7 +98,7 @@ public class StoreOrderInfoResponse implements Serializable {
     private BigDecimal refundPrice;
 
     @ApiModelProperty(value = "使用积分")
-    private Integer useIntegral;
+    private BigDecimal useIntegral;
 
     @ApiModelProperty(value = "给用户退了多少积分")
     private Integer backIntegral;
