+++ crmeb-common/src/main/java/com/cshy/common/model/response/UserCenterResponse.java
@@ -40,7 +40,7 @@ public class UserCenterResponse implements Serializable {
     private BigDecimal nowMoney;
 
     @ApiModelProperty(value = "用户剩余积分")
-    private Integer integral;
+    private BigDecimal integral;
 
     @ApiModelProperty(value = "用户剩余经验")
     private Integer experience;
