+++ crmeb-service/src/main/java/com/cshy/service/service/user/UserService.java
@@ -156,7 +156,7 @@ public interface UserService extends IService<User> {
      * @param type 增加add、扣减sub
      * @return 更新后的用户对象
      */
-    Boolean updateIntegral(User user, Integer integral, String type);
+    Boolean updateIntegral(User user, BigDecimal integral, String type);
 
     /**
      * 手机号注册用户
@@ -216,7 +216,7 @@ public interface UserService extends IService<User> {
      * @param nowIntegral 历史积分
      * @param type 类型：add—添加，sub—扣减
      */
-    Boolean operationIntegral(Integer uid, Integer integral, Integer nowIntegral, String type);
+    Boolean operationIntegral(Integer uid, BigDecimal integral, BigDecimal nowIntegral, String type);
 
     /**
      * PC后台分销员列表
